[KAIZEN-0] fjerne common modul
ikke lengre ibruk etter sletting av login-app
