--- common/build.gradle.kts
@@ -1,57 +1,0 @@
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-
-val ktor_version: String by project
-val logback_version: String by project
-val logstash_version: String by project
-val prometheus_version: String by project
-val modia_common_version: String by project
-val junit_version: String by project
-val unleash_version: String by project
-
-plugins {
-    `java-library`
-    id("setup.repository")
-    kotlin("jvm") version "1.6.21"
-    kotlin("plugin.serialization") version "1.6.21"
-}
-
-dependencies {
-    implementation(kotlin("stdlib"))
-    implementation("io.ktor:ktor-server-netty:$ktor_version")
-    implementation("io.ktor:ktor-server-status-pages:$ktor_version")
-    implementation("io.ktor:ktor-server-content-negotiation:$ktor_version")
-    implementation("io.ktor:ktor-server-forwarded-header:$ktor_version")
-    implementation("io.ktor:ktor-server-call-logging:$ktor_version")
-    implementation("io.ktor:ktor-server-call-id:$ktor_version")
-    implementation("io.ktor:ktor-server-metrics-micrometer:$ktor_version")
-    implementation("io.ktor:ktor-server-auth:$ktor_version")
-    implementation("io.ktor:ktor-serialization:$ktor_version")
-    implementation("io.ktor:ktor-serialization-kotlinx-json:$ktor_version")
-    implementation("io.ktor:ktor-client:$ktor_version")
-    implementation("io.ktor:ktor-client-cio:$ktor_version")
-    implementation("io.ktor:ktor-client-auth:$ktor_version")
-    implementation("io.ktor:ktor-client-logging:$ktor_version")
-    implementation("io.ktor:ktor-client-content-negotiation:$ktor_version")
-    implementation("ch.qos.logback:logback-classic:$logback_version")
-    implementation("net.logstash.logback:logstash-logback-encoder:$logstash_version")
-    implementation("io.micrometer:micrometer-registry-prometheus:$prometheus_version")
-    implementation("no.nav.personoversikt:crypto:$modia_common_version")
-    implementation("io.getunleash:unleash-client-java:$unleash_version")
-
-    testImplementation("org.junit.jupiter:junit-jupiter:$junit_version")
-}
-
-java {
-    sourceCompatibility = JavaVersion.VERSION_11
-    targetCompatibility = JavaVersion.VERSION_11
-}
-tasks.withType<KotlinCompile> {
-    kotlinOptions.jvmTarget = "11"
-}
-
-tasks.test {
-    useJUnitPlatform()
-    testLogging {
-        events("passed", "skipped", "failed")
-    }
-}
