+++ frontend-app/build.gradle.kts
@@ -19,28 +19,30 @@ plugins {
 
 dependencies {
     implementation(kotlin("stdlib"))
-    implementation(project(":common"))
     implementation("io.ktor:ktor-server-netty:$ktor_version")
     implementation("io.ktor:ktor-server-default-headers:$ktor_version")
-    implementation("io.ktor:ktor-server-auth:$ktor_version")
     implementation("io.ktor:ktor-server-auth-jwt:$ktor_version")
     implementation("io.ktor:ktor-server-sessions:$ktor_version")
     implementation("io.ktor:ktor-server-status-pages:$ktor_version")
+    implementation("io.ktor:ktor-server-content-negotiation:$ktor_version")
+    implementation("io.ktor:ktor-server-forwarded-header:$ktor_version")
+    implementation("io.ktor:ktor-server-call-logging:$ktor_version")
+    implementation("io.ktor:ktor-server-call-id:$ktor_version")
+    implementation("io.ktor:ktor-server-metrics-micrometer:$ktor_version")
     implementation("io.ktor:ktor-client:$ktor_version")
+    implementation("io.ktor:ktor-client-logging:$ktor_version")
+    implementation("io.ktor:ktor-client-auth:$ktor_version")
     implementation("io.ktor:ktor-client-apache:$ktor_version")
     implementation("io.ktor:ktor-client-content-negotiation:$ktor_version")
     implementation("io.ktor:ktor-serialization:$ktor_version")
     implementation("io.ktor:ktor-serialization-kotlinx-json:$ktor_version")
     implementation("redis.clients:jedis:$jedis_version")
-    implementation("ch.qos.logback:logback-classic:$logback_version")
-    implementation("net.logstash.logback:logstash-logback-encoder:$logstash_version")
     implementation("no.nav.common:token-client:$java_common_version")
     implementation("no.nav.personoversikt:crypto:$modia_common_version")
     implementation("io.micrometer:micrometer-registry-prometheus:$prometheus_version")
     implementation("io.getunleash:unleash-client-java:$unleash_version")
-    implementation("io.ktor:ktor-server-auth-jvm:2.0.2")
-    implementation("io.ktor:ktor-server-core-jvm:2.0.2")
-    implementation("io.ktor:ktor-server-sessions-jvm:2.0.2")
+    implementation("ch.qos.logback:logback-classic:$logback_version")
+    implementation("net.logstash.logback:logstash-logback-encoder:$logstash_version")
 
     testImplementation("org.junit.jupiter:junit-jupiter:$junit_version")
 }
