+++ frontend-app/src/main/kotlin/no/nav/modialogin/Main.kt
@@ -3,10 +3,10 @@ package no.nav.modialogin
 import io.ktor.server.application.*
 import io.ktor.server.plugins.defaultheaders.*
 import io.ktor.server.request.*
-import no.nav.modialogin.common.KtorServer.log
-import no.nav.modialogin.common.KtorServer.server
-import no.nav.modialogin.common.features.DefaultFeatures.installDefaultFeatures
-import no.nav.modialogin.common.features.installNaisFeature
+import no.nav.modialogin.utils.KtorServer.log
+import no.nav.modialogin.utils.KtorServer.server
+import no.nav.modialogin.features.DefaultFeatures.installDefaultFeatures
+import no.nav.modialogin.features.installNaisFeature
 import no.nav.modialogin.features.HostStaticFilesFeature
 import no.nav.modialogin.features.HostStaticFilesFeature.Companion.installHostStaticFilesFeature
 import no.nav.modialogin.features.ReferrerPolicyFeature.applyReferrerPolicyFeature
