+++ frontend-app/src/main/kotlin/no/nav/modialogin/features/authfeature/SessionCache.kt
@@ -7,7 +7,6 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.runBlocking
 import kotlinx.serialization.builtins.serializer
 import no.nav.modialogin.RedisConfig
-import no.nav.modialogin.auth.OidcClient
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.minutes
