+++ frontend-app/src/main/kotlin/no/nav/modialogin/features/bffproxyfeature/directives/RespondDirectiveSpecification.kt
@@ -3,8 +3,8 @@ package no.nav.modialogin.features.bffproxyfeature.directives
 import io.ktor.http.*
 import io.ktor.server.application.*
 import io.ktor.server.response.*
-import no.nav.modialogin.common.KotlinUtils.cutoff
-import no.nav.modialogin.common.Templating
+import no.nav.modialogin.utils.KotlinUtils.cutoff
+import no.nav.modialogin.utils.Templating
 import no.nav.modialogin.features.bffproxyfeature.BFFProxy
 import no.nav.modialogin.features.bffproxyfeature.ResponseDirectiveHandler
