+++ frontend-app/src/main/kotlin/no/nav/modialogin/utils/UnleashTemplateSource.kt
@@ -1,10 +1,9 @@
-package no.nav.modialogin.features
+package no.nav.modialogin.utils
 
 import io.getunleash.Unleash
 import io.getunleash.UnleashContext
 import io.ktor.server.application.*
 import io.ktor.server.auth.*
-import no.nav.modialogin.common.TemplatingEngine
 import no.nav.modialogin.features.authfeature.TokenPrincipal
 
 object UnleashTemplateSource {
