+++ frontend-app/src/main/resources/common/logging/logback-securelogs.xml
