+++ frontend-app/src/main/resources/common/logging/logback-stdout-json.xml
