+++ settings.gradle.kts
@@ -1,4 +1,3 @@
 rootProject.name = "modialogin"
-include("common")
 include("frontend-app")
 include("oidc-stub")
